Import gcc-3.0-mint-4.diff from gcc-3.0-mint-4.tar.gz
Source:
http://patrice.mandin.pagesperso-orange.fr/v3/en/patch-utils.html

Gcc-3.0 for Mint				July, 31th 2001

Here is my fourth patch for gcc-3.0.

When cross-compiling, the 'specs' file should be filled with the target as and
ld, but it is not the case (and so I manually modify it to continue
compilation). So when you have an error, edit the specs file,
and change ld and as with m68k-atari-mint-ld and m68k-atari-mint-as.

I modified configure.in in the directory libstdc++-v3, so you must run
'autoconf' before build.

You need to patch the mintlib, because a function 'islower' is needed.
Maybe some others are needed also.

I modified it to be able to compile position independant code (added '-fPIC
switch), first step for a dynamic linker. But it needs another object file
format for the final library/executable, maybe ELF ?

--
Patrice Mandin <pmandin@caramail.com>
http://www.multimania.com/pmandin