Import gcc-3.0.1-mint.diff from gcc-3.0.1-mint.tar.gz
Gcc-3.0.1 for Mint				August, 31th 2001

Here is my patch for gcc-3.0.1.

When cross-compiling, the 'specs' file should be filled with the target as and
ld, but it is not the case (and so I manually modify it to continue
compilation). So when you have an error, edit the specs file,
and change ld and as with m68k-atari-mint-ld and m68k-atari-mint-as.

I modified configure.in in the directory libstdc++-v3, so you must run
'autoconf' before build.

I modified it to be able to compile position independant code (added '-fPIC
switch), first step for a dynamic linker. But it needs another object file
format for the final library/executable, maybe ELF ?

--
Patrice Mandin <pmandin@caramail.com>
http://www.multimania.com/pmandin