Import gcc-4.2.0-mint-20070616.diff
The jmp instruction used in ASM_RETURN_CASE_JUMP expects that the array
of pointers is located exactly after it. So we force a zero alignment
in ADDR_VEC_ALIGN().

Use .balign to align on a number of bytes, as .align may behave like .p2align
depending on the gas configutation.