Import gcc-4.2.2-mint-20080107.diff
Fixed a bug in ld: CONSTRUCTORS (the list of the global
constructors and destructors) is now relocated as expected.

Warning! The previous versions of libgcc contained a patch
as a workaround to this bug. Since the bug is fixed, I removed
the libgcc patch. That means that the new binutils are incompatible
with the old GCC, and the new GCC is incompatible with the old binutils
(only for programs using constructors for global variables)!
Global contructors worked before. They still work now.
But remember to update the binutils and GCC packages at the same time.
Matching binutils patch is binutils-2.18-mint-20080106.diff.