Import gcc-4.2.2-mint-20080127.patch
All the tools now display the MiNT patch date with --version.

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=34947
Due to this bug, which was especially causing Quake to crash
when compiled with full optimization, I removed the automatic
implicit use of -fomit-frame-pointer when compiling with -O.
Thanks to Miro Kropáček for reporting this bug.

Added the necessary patches do build natively GCC on MiNT
with GCC 2.95.3.
Once again, thanks to Miro Kropáček for providing these patches.