Import gcc-4.2.3-mint-20080223.patch
Fixed the Quake bug. Now programs can use -m68020-60, -O
and -fomit-frame-pointer at the same time.
Note that if the option -fomit-frame-pointer is not explicitly specified,
the frame pointer will always be used, regardless to the optimization level.
Thanks to Richard Sandiford for having made this patch, and to Miro Kropáček
for having found it.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=34947