Import gcc-4.5.2-mint-20110128.patch
On the -m68020-60 configuration, floats are now returned into d0/d1
instead of fp0. The ABI is now compatible between all the configurations,
and the libraries can be mixed.

Added the nonstandard define __M68881__ whenever __HAVE_68881__ is defined.
This means a 680x0 FPU (not ColdFire) is available and should be used.
This new define is for the compatibility with older sources.
New sources should only test __HAVE_68881__.