Import gcc-4.5.2-mint-20110206.patch
Never automatically link with the debug or profile standard libraries.
To do so, compile them yourself then add -lc_g or -lc_p on the command
line. As a consequence, linking with -g is now safe in any case.

Fixed GCC bug 47672 about compilation of math-68881.h with -std=c99
or -std=gnu99. Thanks to Paweł Góralski to have reported this problem.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47672